Move all query definitions to queries file
